refacorisation pour avoir le meme Model, dossier joueurLib supprimé et le fichier functions.php refacoriser dans lib, dossier joueur deplacer dans vue
