+++ Pays.php
@@ -105,4 +105,3 @@ function createOrUpdate() {
     }
 
 }
-?>
