+++ class/player/Player.php
@@ -1,6 +1,5 @@
 <?php
 
-
 require_once dirname(__DIR__) . "/model/Model.php";
 
 /**
