+++ vue/joueur/delete.php
@@ -3,8 +3,9 @@
 
 $delete = null;
 
-require_once ("../class/player/Player.php");
-require_once ("../joueurLib/functions.php");
+
+require_once ("../../class/player/Player.php");
+require_once ("../../lib/functions.php");
 
 $delete = new Player;
